feat(linux): Add Liberation font

+++ linux/build.sh
@@ -33,7 +33,6 @@ echo "*                                                                    *"
 echo "**********************************************************************"
 
 # Configuration
-ULTRASCHALL_THREAD_COUNT=$(sysctl -n hw.ncpu)
 ULTRASCHALL_CMAKE_TOOL=cmake
 ULTRASCHALL_PANDOC_TOOL=pandoc
 
@@ -46,13 +45,20 @@ ULTRASCHALL_BUILD_RELEASE=0
 
 ULTRASCHALL_ROOT_DIRECTORY=$(pwd)
 ULTRASCHALL_BUILD_DIRECTORY="$ULTRASCHALL_ROOT_DIRECTORY/build"
+ULTRASCHALL_BUILD_LOG="$ULTRASCHALL_ROOT_DIRECTORY/build.log"
 
 ULTRASCHALL_PLUGIN_URL="https://github.com/Ultraschall/ultraschall-plugin.git"
+ULTRASCHALL_PLUGIN_BRANCH="main"
+
 ULTRASCHALL_SOUNDBOARD_URL="https://github.com/Ultraschall/ultraschall-soundboard.git"
+ULTRASCHALL_SOUNDBOARD_BRANCH="main"
+
 # ULTRASCHALL_PORTABLE_URL="https://github.com/Ultraschall/ultraschall-portable.git"
 ULTRASCHALL_PORTABLE_URL="https://github.com/nethad/ultraschall-portable.git"
 ULTRASCHALL_PORTABLE_BRANCH="linux-project-and-track-templates"
+
 ULTRASCHALL_ASSETS_URL="https://github.com/Ultraschall/ultraschall-assets.git"
+ULTRASCHALL_ASSETS_BRANCH="master"
 
 STUDIO_LINK_PLUGIN_RELEASE=v21.07.0-stable
 
@@ -63,15 +69,17 @@ elif [ "$1" = "--release" ]; then
   ULTRASCHALL_BUILD_RELEASE=1
 fi
 
-echo "Checking whether required software packages are installed..."
-ULTRASCHALL_DEPENDENCIES="cmake pandoc ninja-build libxrandr-dev libxinerama-dev libxcursor-dev libasound2-dev libcurl4-openssl-dev"
-dpkg -l $ULTRASCHALL_DEPENDENCIES > /dev/null
-if [ $? -ne 0 ]; then
-  echo "Dependency check failed, required packages:"
-  echo "$ULTRASCHALL_DEPENDENCIES"
-  exit -1
-fi
-echo "Done."
+# ULTRASCHALL_DEPENDENCIES="cmake pandoc ninja-build freetype-devel libXrandr-devel libXinerama-devel libXcursor-devel alsa-lib-devel libcurl-devel"
+
+# echo "Checking whether required software packages are installed..."
+# ULTRASCHALL_DEPENDENCIES="cmake pandoc ninja-build libxrandr-dev libxinerama-dev libxcursor-dev libasound2-dev libcurl4-openssl-dev"
+# dpkg -l $ULTRASCHALL_DEPENDENCIES > /dev/null
+# if [ $? -ne 0 ]; then
+#   echo "Dependency check failed, required packages:"
+#   echo "$ULTRASCHALL_DEPENDENCIES"
+#   exit -1
+# fi
+# echo "Done."
 
 SOURCE_BRANCH='develop'
 if [ $ULTRASCHALL_BUILD_RELEASE = 1 ]; then
@@ -89,7 +97,13 @@ fi
 
 ULTRASCHALL_INSTALLER_DIR="$ULTRASCHALL_BUILD_ID"
 
-# Create folder for intermediate data
+# Create build log
+if [ -f $ULTRASCHALL_BUILD_LOG ]; then
+  rm -f $ULTRASCHALL_BUILD_LOG
+fi
+touch $ULTRASCHALL_BUILD_LOG
+
+# Create intermediate data folder
 if [ ! -d $ULTRASCHALL_BUILD_DIRECTORY ]; then
   mkdir $ULTRASCHALL_BUILD_DIRECTORY
 fi
@@ -98,7 +112,7 @@ pushd $ULTRASCHALL_BUILD_DIRECTORY > /dev/null
 
 if [ ! -d ultraschall-plugin ]; then
   echo "Downloading Ultraschall REAPER Plug-in..."
-  git clone --branch $SOURCE_BRANCH --depth 1 $ULTRASCHALL_PLUGIN_URL ultraschall-plugin
+  git clone --branch $ULTRASCHALL_PLUGIN_BRANCH --depth 1 $ULTRASCHALL_PLUGIN_URL ultraschall-plugin >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
   if [ ! -d ultraschall-plugin ]; then
     echo "Failed to download Ultraschall REAPER Plug-in."
     exit -1
@@ -106,14 +120,14 @@ if [ ! -d ultraschall-plugin ]; then
 else
   echo "Updating Ultraschall REAPER Plug-in..."
   pushd ultraschall-plugin > /dev/null
-  git pull
+  git pull >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
   popd > /dev/null
 fi
 echo "Done."
 
 if [ ! -d ultraschall-soundboard ]; then
   echo "Downloading Ultraschall Soundboard..."
-  git clone --branch main --depth 1 $ULTRASCHALL_SOUNDBOARD_URL ultraschall-soundboard
+  git clone --branch $ULTRASCHALL_SOUNDBOARD_BRANCH --depth 1 $ULTRASCHALL_SOUNDBOARD_URL ultraschall-soundboard >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
   if [ ! -d ultraschall-soundboard ]; then
     echo "Failed to download Ultraschall Soundboard."
     exit -1
@@ -121,14 +135,14 @@ if [ ! -d ultraschall-soundboard ]; then
 else
   echo "Updating Ultraschall Soundboard..."
   pushd ultraschall-soundboard > /dev/null
-  git pull
+  git pull >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
   popd > /dev/null
 fi
 echo "Done."
 
 if [ ! -d ultraschall-portable ]; then
   echo "Downloading Ultraschall REAPER API..."
-  git clone --branch $ULTRASCHALL_PORTABLE_BRANCH --depth 1 $ULTRASCHALL_PORTABLE_URL ultraschall-portable
+  git clone --branch $ULTRASCHALL_PORTABLE_BRANCH --depth 1 $ULTRASCHALL_PORTABLE_URL ultraschall-portable >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
   if [ ! -d ultraschall-portable ]; then
     echo "Failed to download Ultraschall REAPER API."
     exit -1
@@ -136,14 +150,14 @@ if [ ! -d ultraschall-portable ]; then
 else
   echo "Updating Ultraschall REAPER API..."
   pushd ultraschall-portable > /dev/null
-  git pull
+  git pull >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
   popd > /dev/null
 fi
 echo "Done."
 
 if [ ! -d ultraschall-assets ]; then
   echo "Downloading Ultraschall REAPER Resources..."
-  git clone $ULTRASCHALL_ASSETS_URL ultraschall-assets
+  git clone --branch $ULTRASCHALL_ASSETS_BRANCH --depth 1 $ULTRASCHALL_ASSETS_URL ultraschall-assets >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
   if [ ! -d ultraschall-assets ]; then
     echo "Failed to download Ultraschall REAPER Resources."
     exit -1
@@ -151,7 +165,7 @@ if [ ! -d ultraschall-assets ]; then
 else
   echo "Updating Ultraschall REAPER Resources..."
   pushd ultraschall-assets > /dev/null
-  git pull
+  git pull >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
   popd > /dev/null
 fi
 echo "Done."
@@ -176,6 +190,9 @@ fi
 if [ ! -d custom-plugins ]; then
   mkdir custom-plugins
 fi
+if [ ! -d fonts ]; then
+  mkdir fonts
+fi
 popd > /dev/null
 echo "Done."
 
@@ -186,14 +203,16 @@ $ULTRASCHALL_PANDOC_TOOL --from=markdown --to=html --embed-resources --standalon
 echo "Done."
 
 pushd ultraschall-plugin > /dev/null
-echo "Building Ultraschall REAPER Plug-in..."
-$ULTRASCHALL_CMAKE_TOOL -B build -G Ninja -DCMAKE_BUILD_TYPE=Release --log-level=ERROR 2> build.log
+echo "Configuring Ultraschall REAPER Plug-in..."
+$ULTRASCHALL_CMAKE_TOOL -Bbuild -GNinja -Wno-dev -DCMAKE_BUILD_TYPE=Release --log-level=ERROR >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
 if [ $? -ne 0 ]; then
   cat build.log
   echo "Failed to configure Ultraschall REAPER Plug-in."
   exit -1
 fi
-$ULTRASCHALL_CMAKE_TOOL --build build --target reaper_ultraschall --config Release -j $ULTRASCHALL_THREAD_COUNT 2>> build.log
+echo "Done."
+echo "Building Ultraschall REAPER Plug-in..."
+$ULTRASCHALL_CMAKE_TOOL --build build --target reaper_ultraschall --config Release -j >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
 if [ $? -ne 0 ]; then
   cat build.log
   echo "Failed to build Ultraschall REAPER Plug-in."
@@ -204,7 +223,7 @@ echo "Done."
 
 pushd ultraschall-soundboard > /dev/null
 echo "Configuring Ultraschall Soundboard..."
-$ULTRASCHALL_CMAKE_TOOL -B build -G Ninja -DCMAKE_BUILD_TYPE=Release --log-level=ERROR 2> build.log
+$ULTRASCHALL_CMAKE_TOOL -Bbuild -GNinja -Wno-dev -DCMAKE_BUILD_TYPE=Release --log-level=ERROR >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
 if [ $? -ne 0 ]; then
   cat build.log
   echo "Failed to configure Ultraschall Soundboard."
@@ -212,7 +231,7 @@ if [ $? -ne 0 ]; then
 fi
 echo "Done."
 echo "Building Ultraschall Soundboard..."
-$ULTRASCHALL_CMAKE_TOOL --build build --target UltraschallSoundboard_VST3 --config Release -j $ULTRASCHALL_THREAD_COUNT 2>> build.log
+$ULTRASCHALL_CMAKE_TOOL --build build --target UltraschallSoundboard_VST3 --config Release -j >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
 if [ $? -ne 0 ]; then
   cat build.log
   echo "Failed to build Ultraschall Soundboard."
@@ -235,15 +254,51 @@ cp ultraschall-portable/UserPlugins/ultraschall_api.lua ultraschall-api/
 cp ultraschall-portable/UserPlugins/ultraschall_api_readme.txt ultraschall-api/
 echo "Done."
 
-echo "Fetching StudioLink plugin"
-curl https://download.studio.link/releases/$STUDIO_LINK_PLUGIN_RELEASE/linux/vst/studio-link-plugin.zip -o studio-link-plugin.zip
+echo "Downloading StudioLink plugin..."
+curl -LO https://download.studio.link/releases/$STUDIO_LINK_PLUGIN_RELEASE/linux/vst/studio-link-plugin.zip >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
+if [ $? -ne 0 ]; then
+  echo "Failed to download StudioLink plugin."
+  exit -1
+fi
 rm -rf studio-link-plugin.vst
-unzip -d studio-link-plugin.vst studio-link-plugin.zip
+unzip -d studio-link-plugin.vst studio-link-plugin.zip >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
+if [ $? -ne 0 ]; then
+  echo "Failed to extract StudioLink plugin."
+  exit -1
+fi
+echo "Done."
 
-echo "Fetching StudioLink OnAir plugin"
-curl https://download.studio.link/releases/$STUDIO_LINK_PLUGIN_RELEASE/linux/studio-link-plugin-onair.zip -o studio-link-plugin-onair.zip
+echo "Downloading StudioLink OnAir plugin..."
+curl -LO https://download.studio.link/releases/$STUDIO_LINK_PLUGIN_RELEASE/linux/studio-link-plugin-onair.zip >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
+if [ $? -ne 0 ]; then
+  echo "Failed to download StudioLink OnAir plugin."
+  exit -1
+fi
 rm -rf studio-link-onair.lv2
-unzip studio-link-plugin-onair.zip
+unzip studio-link-plugin-onair.zip >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
+if [ $? -ne 0 ]; then
+  echo "Failed to extract StudioLink OnAir plugin."
+  exit -1
+fi
+echo "Done."
+
+echo "Downloading Liberation fonts..."
+curl -LO "https://github.com/liberationfonts/liberation-fonts/files/7261482/liberation-fonts-ttf-2.1.5.tar.gz" >> $ULTRASCHALL_BUILD_LOG 2>> $ULTRASCHALL_BUILD_LOG
+if [ $? -ne 0 ]; then
+  echo "Failed to download Liberation fonts."
+  exit -1
+fi
+rm -rf liberation-fonts-ttf-2.1.5
+tar -xf liberation-fonts-ttf-2.1.5.tar.gz
+if [ $? -ne 0 ]; then
+  echo "Failed to extract Liberation fonts."
+  exit -1
+fi
+echo "Done."
+
+echo "Copying Liberation fonts..."
+cp liberation-fonts-ttf-2.1.5/*.ttf "$ULTRASCHALL_INSTALLER_DIR/fonts"
+echo "Done."
 
 echo "Copying Ultraschall resources..."
 cp ultraschall-assets/keyboard-layout/Keymap.pdf "$ULTRASCHALL_INSTALLER_DIR/resources/Ultraschall Keyboard Layout.pdf"
